# GH-115060: Speed up `pathlib.Path.glob()` by omitting initial `stat()`