--- a/Doc/library/pathlib.rst
+++ b/Doc/library/pathlib.rst
@@ -1004,7 +1004,8 @@ call fails (for example because the path doesn't exist).
    .. seealso::
       :ref:`pathlib-pattern-language` documentation.
 
-   This method suppresses any :exc:`OSError` exceptions raised while scanning.
+   This method suppresses any :exc:`OSError` exceptions that are raised while
+   scanning the filesystem.
 
    By default, or when the *case_sensitive* keyword-only argument is set to
    ``None``, this method matches paths using platform-specific casing rules: