--- a/Lib/pathlib/__init__.py
+++ b/Lib/pathlib/__init__.py
@@ -607,11 +607,9 @@ class Path(_abc.PathBase, PurePath):
         if raw[-1] in (self.parser.sep, self.parser.altsep):
             # GH-65238: pathlib doesn't preserve trailing slash. Add it back.
             parts.append('')
-        if not self.is_dir():
-            return iter([])
         select = self._glob_selector(parts[::-1], case_sensitive, recurse_symlinks)
         root = str(self)
-        paths = select(root, exists=True)
+        paths = select(root)
 
         # Normalize results
         if root == '.':