--- a/Misc/NEWS.d/next/Library/2024-04-13-01-45-15.gh-issue-115060.IxoM03.rst
+++ b/Misc/NEWS.d/next/Library/2024-04-13-01-45-15.gh-issue-115060.IxoM03.rst
@@ -1,3 +1,3 @@
-Speed up :meth:`pathlib.Path.glob` my omitting an initial
+Speed up :meth:`pathlib.Path.glob` by omitting an initial
 :meth:`~pathlib.Path.is_dir` call. As a result of this change,
 :meth:`~pathlib.Path.glob` can no longer raise :exc:`OSError`.