@barneygale 2024-04-13
Thanks! I think it's probably not important enough for `.. versionchanged::`, particularly as we don't document the sorts of `OSError` that are raised or suppressed from `is_dir()`.