@barneygale 2024-04-13
On reflection, I think this works best as a `.. versionchanged::` directive.